Made container creation not DIAF if invalid code is committed.
Due to how container startup happens, if there's a syntax error in some python files the container will DIAF on startup and lead to the Codespace starting up in safe mode.  A safe mode container can not push code to a repo, so a student can get themselves into a situation where they can not fix their problems in Codespaces.  This patch prevents that situation from happening.